Disable Html for markdown
This disables html for the markdown editor so users and sellers are not able to write their custom html inside the editor

This has a problem. All the products and terms and conditions that were created with the old markdown editor will have to be converted from html to markdown. You can easily convert them using an online converter but it has to be done manually by the seller